# WIP Use --consumer-only and --no-coverage in certain Cirrus tests to speed up setup