--- a/.cirrus.yml
+++ b/.cirrus.yml
@@ -260,7 +260,7 @@ task:
   setup_script:
     - flutter config --no-analytics
     - flutter doctor -v
-    - flutter update-packages
+    - flutter update-packages --no-coverage
     - git fetch origin master
   test_all_script:
     - dart --enable-asserts dev\bots\test.dart
@@ -314,7 +314,7 @@ task:
   setup_script:
     - flutter config --no-analytics
     - flutter doctor -v
-    - flutter update-packages
+    - flutter update-packages --no-coverage
     - git fetch origin master
   matrix:
     - name: tests_widgets-windows
@@ -415,7 +415,7 @@ task:
     - git fetch origin master # To set FETCH_HEAD
   setup_script:
     - bin/flutter config --no-analytics
-    - bin/flutter update-packages
+    - bin/flutter update-packages --no-coverage
   test_all_script:
     - ./dev/bots/deploy_gallery.sh
 
@@ -437,7 +437,7 @@ task:
   setup_script:
     - bin/flutter config --no-analytics
     - bin/flutter doctor -v
-    - bin/flutter update-packages
+    - bin/flutter update-packages --no-coverage
   test_all_script:
     - ulimit -S -n 2048 # https://github.com/flutter/flutter/issues/2976
     - bin/cache/dart-sdk/bin/dart --enable-asserts dev/bots/test.dart
@@ -488,7 +488,7 @@ task:
   setup_script:
     - bin/flutter config --no-analytics
     - bin/flutter doctor -v
-    - bin/flutter update-packages
+    - bin/flutter update-packages --consumer-only --no-coverage
   matrix:
     - name: integration_tests-macos
       only_if: $CIRRUS_BRANCH == 'master'