--- a/.cirrus.yml
+++ b/.cirrus.yml
@@ -256,7 +256,8 @@ task:
   setup_script:
     - flutter config --no-analytics
     - flutter doctor -v
-    - flutter update-packages --no-coverage
+    - bin/flutter update-packages --consumer-only --no-coverage
+    - bin/flutter packages get
     - git fetch origin master
   test_all_script:
     - dart --enable-asserts dev\bots\test.dart
@@ -307,7 +308,8 @@ task:
   setup_script:
     - flutter config --no-analytics
     - flutter doctor -v
-    - flutter update-packages --no-coverage
+    - bin/flutter update-packages --consumer-only --no-coverage
+    - bin/flutter packages get
     - git fetch origin master
   matrix:
     - name: tests_widgets-windows
@@ -408,7 +410,8 @@ task:
     - git fetch origin master # To set FETCH_HEAD
   setup_script:
     - bin/flutter config --no-analytics
-    - bin/flutter update-packages --no-coverage
+    - bin/flutter update-packages --consumer-only --no-coverage
+    - bin/flutter packages get
   test_all_script:
     - ./dev/bots/deploy_gallery.sh
 
@@ -430,7 +433,8 @@ task:
   setup_script:
     - bin/flutter config --no-analytics
     - bin/flutter doctor -v
-    - bin/flutter update-packages --no-coverage
+    - bin/flutter update-packages --consumer-only --no-coverage
+    - bin/flutter packages get
   test_all_script:
     - ulimit -S -n 2048 # https://github.com/flutter/flutter/issues/2976
     - bin/cache/dart-sdk/bin/dart --enable-asserts dev/bots/test.dart
@@ -480,6 +484,7 @@ task:
     - bin/flutter config --no-analytics
     - bin/flutter doctor -v
     - bin/flutter update-packages --consumer-only --no-coverage
+    - bin/flutter packages get
   matrix:
     - name: integration_tests-macos
       only_if: $CIRRUS_BRANCH == 'master'