## Description
1. Skip downloading lcov coverage data in Cirrus tests that don't need it.
2. Only download consumer-facing packages for integration tests.

```
flutter update-packages  26.48s user 5.66s system 164% cpu 19.583 total
flutter update-packages --consumer-only --no-coverage  2.62s user 0.70s system 155% cpu 2.135 total
```