@jmagman 2019-08-10
--consumer-only didn't do what I thought it did, and the Cirrus working directory wasn't a dev directory like I thought it was, so `flutter packages get` didn't work.